Rename getGroupMessagesAction to groupMessagesAction. Returns only relevant details of logged in user to the store and refactor code to reflect changes
