+++ app/client/actions/setSelectedGroupAction.js
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import isEmpty from 'lodash/isEmpty';
 import { SET_SELECTED_GROUP } from '../actions/types';
 
 export function setSelectedGroup(group) {
@@ -8,22 +7,3 @@ export function setSelectedGroup(group) {
     selectedGroup: group
   };
 }
-
-// export function setSelectedGroup(group) {
-//   if (isEmpty(group)) {
-//     return (dispatch) => {
-//       dispatch(setSelectedGroupMsgs({}));
-//     };
-//   }
-
-//   const groupId = group.id;
-//   const request = axios.get(`/api/group/${groupId}/messages`);
-
-//   return (dispatch) => {
-//     return request.then((res) => {
-//       const messages = res.data;
-//       // const messagesObj = mapKeys(res.data, 'message_id');
-//       dispatch(setSelectedGroupMsgs({ id: group.id, name: group.name, messages }));
-//     });
-//   };
-// }
