+++ app/client/actions/types.js
@@ -5,3 +5,4 @@ export const SET_USER_GROUPS = 'SET_USER_GROUPS';
 export const SET_SELECTED_GROUP = 'SET_SELECTED_GROUP';
 export const SET_GROUP_MESSAGES = 'SET_GROUP_MESSAGES';
 export const POST_NEW_MESSAGE = 'POST_NEW_MESSAGE';
+export const DELETE_CURRENT_USER = 'DELETE_CURRENT_USER';
