+++ app/client/reducers/reducer_authenticatedUser.js
@@ -1,5 +1,5 @@
 import isEmpty from 'lodash/isEmpty';
-import { SET_CURRENT_USER } from '../actions/types';
+import { SET_CURRENT_USER, DELETE_CURRENT_USER } from '../actions/types';
 
 const initialState = {
   isAuthenticated: false,
@@ -11,8 +11,21 @@ export default (state = initialState, action = {}) => {
     case SET_CURRENT_USER:
       return {
         isAuthenticated: !isEmpty(action.user),
-        user: action.user
+        user: {
+          id: action.user.data.id,
+          firstname: action.user.data.firstname,
+          lastname: action.user.data.lastname,
+          username: action.user.data.username,
+          email: action.user.data.email
+        }
       };
+
+      case DELETE_CURRENT_USER:
+      return {
+        isAuthenticated: !isEmpty(action.user),
+        user: action.user
+      }
+
     default:
       return state;
   }
